requirements.txt
dropbox>=11.36.0
pycryptodome>=3.17
discord>=2.2.2
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.9
python-dotenv>=1.0.0
detectlanguage>=1.5.0
langdetect>=1.0.9
deep_translator>=1.10.1
requests>=2.28.2
langcodes>=3.3.0
language_data>=1.1